+++ renderers/react/a2ui_explorer/src/demo-catalog.ts
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2024 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {Catalog} from '@a2ui/web_core/v0_9';
+import {BASIC_FUNCTIONS} from '@a2ui/web_core/v0_9/basic_catalog';
+import {basicCatalog, toWebComponent} from '@a2ui/react/v0_9';
+import {customSliderComponent} from './custom-slider';
+import {customGridComponent} from './custom-grid';
+
+/**
+ * A catalog specific to the React demo explorer, extending the basic catalog
+ * with custom components.
+ */
+export const demoCatalog = new Catalog(
+  basicCatalog.id,
+  [
+    ...basicCatalog.components.values(),
+    toWebComponent(customSliderComponent),
+    toWebComponent(customGridComponent),
+  ],
+  BASIC_FUNCTIONS,
+);